Download snapshot

Downloads a full copy (snapshot) of an R package repository.

Usage

repo_snapshot(repo, destdir = "snapshot", bin_versions = r_version())

r_version()

Arguments

repo

url of the cran-like repository to snapshot

bin_versions

vector with versions of R to download the win/mac binary packages. The default is to download binaries only for your local version. Set to NULL to not download any binaries.

Examples

repo_snapshot('https://jeroen.r-universe.dev', bin_versions = c("4.1", "4.2", "4.3"))
#> Mirroring repo: https://jeroen.r-universe.dev/src/contrib
#> Mirroring repo: https://jeroen.r-universe.dev/bin/windows/contrib/4.1
#> Warning: Repository is empty: https://jeroen.r-universe.dev/bin/windows/contrib/4.1
#> Mirroring repo: https://jeroen.r-universe.dev/bin/macosx/contrib/4.1
#> Warning: Repository is empty: https://jeroen.r-universe.dev/bin/macosx/contrib/4.1
#> Mirroring repo: https://jeroen.r-universe.dev/bin/windows/contrib/4.2
#> Mirroring repo: https://jeroen.r-universe.dev/bin/macosx/contrib/4.2
#> Mirroring repo: https://jeroen.r-universe.dev/bin/windows/contrib/4.3
#> Mirroring repo: https://jeroen.r-universe.dev/bin/macosx/contrib/4.3
#>   [1] "bin/macosx/contrib/4.2/PACKAGES"            
#>   [2] "bin/macosx/contrib/4.2/PACKAGES.gz"         
#>   [3] "bin/macosx/contrib/4.2/RAppArmor_3.2.4.tgz" 
#>   [4] "bin/macosx/contrib/4.2/V8_4.4.2.tgz"        
#>   [5] "bin/macosx/contrib/4.2/badgen_1.0.0.tgz"    
#>   [6] "bin/macosx/contrib/4.2/base64_2.0.1.tgz"    
#>   [7] "bin/macosx/contrib/4.2/bcrypt_1.2.tgz"      
#>   [8] "bin/macosx/contrib/4.2/brotli_1.3.0.tgz"    
#>   [9] "bin/macosx/contrib/4.2/curl_5.2.1.tgz"      
#>  [10] "bin/macosx/contrib/4.2/gpg_1.2.9.tgz"       
#>  [11] "bin/macosx/contrib/4.2/js_1.2.tgz"          
#>  [12] "bin/macosx/contrib/4.2/jsonlite_1.8.8.tgz"  
#>  [13] "bin/macosx/contrib/4.2/maketools_1.3.0.tgz" 
#>  [14] "bin/macosx/contrib/4.2/mongolite_2.7.3.tgz" 
#>  [15] "bin/macosx/contrib/4.2/openssl_2.1.1.tgz"   
#>  [16] "bin/macosx/contrib/4.2/protolite_2.3.0.tgz" 
#>  [17] "bin/macosx/contrib/4.2/rjade_0.1.1.tgz"     
#>  [18] "bin/macosx/contrib/4.2/sys_3.4.2.tgz"       
#>  [19] "bin/macosx/contrib/4.2/timestamp.txt"       
#>  [20] "bin/macosx/contrib/4.2/unix_1.5.7.tgz"      
#>  [21] "bin/macosx/contrib/4.2/webp_1.2.0.tgz"      
#>  [22] "bin/macosx/contrib/4.2/webutils_1.2.0.tgz"  
#>  [23] "bin/macosx/contrib/4.3/PACKAGES"            
#>  [24] "bin/macosx/contrib/4.3/PACKAGES.gz"         
#>  [25] "bin/macosx/contrib/4.3/RAppArmor_3.2.4.tgz" 
#>  [26] "bin/macosx/contrib/4.3/V8_4.4.2.tgz"        
#>  [27] "bin/macosx/contrib/4.3/badgen_1.0.0.tgz"    
#>  [28] "bin/macosx/contrib/4.3/base64_2.0.1.tgz"    
#>  [29] "bin/macosx/contrib/4.3/bcrypt_1.2.tgz"      
#>  [30] "bin/macosx/contrib/4.3/brotli_1.3.0.tgz"    
#>  [31] "bin/macosx/contrib/4.3/curl_5.2.1.tgz"      
#>  [32] "bin/macosx/contrib/4.3/gpg_1.2.9.tgz"       
#>  [33] "bin/macosx/contrib/4.3/js_1.2.tgz"          
#>  [34] "bin/macosx/contrib/4.3/jsonlite_1.8.8.tgz"  
#>  [35] "bin/macosx/contrib/4.3/maketools_1.3.0.tgz" 
#>  [36] "bin/macosx/contrib/4.3/mongolite_2.7.3.tgz" 
#>  [37] "bin/macosx/contrib/4.3/openssl_2.1.1.tgz"   
#>  [38] "bin/macosx/contrib/4.3/protolite_2.3.0.tgz" 
#>  [39] "bin/macosx/contrib/4.3/rjade_0.1.1.tgz"     
#>  [40] "bin/macosx/contrib/4.3/sys_3.4.2.tgz"       
#>  [41] "bin/macosx/contrib/4.3/timestamp.txt"       
#>  [42] "bin/macosx/contrib/4.3/unix_1.5.7.tgz"      
#>  [43] "bin/macosx/contrib/4.3/webp_1.2.0.tgz"      
#>  [44] "bin/macosx/contrib/4.3/webutils_1.2.0.tgz"  
#>  [45] "bin/windows/contrib/4.2/PACKAGES"           
#>  [46] "bin/windows/contrib/4.2/PACKAGES.gz"        
#>  [47] "bin/windows/contrib/4.2/V8_4.4.2.zip"       
#>  [48] "bin/windows/contrib/4.2/badgen_1.0.0.zip"   
#>  [49] "bin/windows/contrib/4.2/base64_2.0.1.zip"   
#>  [50] "bin/windows/contrib/4.2/bcrypt_1.2.zip"     
#>  [51] "bin/windows/contrib/4.2/brotli_1.3.0.zip"   
#>  [52] "bin/windows/contrib/4.2/curl_5.2.1.zip"     
#>  [53] "bin/windows/contrib/4.2/gpg_1.2.9.zip"      
#>  [54] "bin/windows/contrib/4.2/js_1.2.zip"         
#>  [55] "bin/windows/contrib/4.2/jsonlite_1.8.8.zip" 
#>  [56] "bin/windows/contrib/4.2/maketools_1.3.0.zip"
#>  [57] "bin/windows/contrib/4.2/mongolite_2.7.3.zip"
#>  [58] "bin/windows/contrib/4.2/openssl_2.1.1.zip"  
#>  [59] "bin/windows/contrib/4.2/protolite_2.3.0.zip"
#>  [60] "bin/windows/contrib/4.2/rjade_0.1.1.zip"    
#>  [61] "bin/windows/contrib/4.2/sys_3.4.2.zip"      
#>  [62] "bin/windows/contrib/4.2/timestamp.txt"      
#>  [63] "bin/windows/contrib/4.2/webp_1.2.0.zip"     
#>  [64] "bin/windows/contrib/4.2/webutils_1.2.0.zip" 
#>  [65] "bin/windows/contrib/4.3/PACKAGES"           
#>  [66] "bin/windows/contrib/4.3/PACKAGES.gz"        
#>  [67] "bin/windows/contrib/4.3/V8_4.4.2.zip"       
#>  [68] "bin/windows/contrib/4.3/badgen_1.0.0.zip"   
#>  [69] "bin/windows/contrib/4.3/base64_2.0.1.zip"   
#>  [70] "bin/windows/contrib/4.3/bcrypt_1.2.zip"     
#>  [71] "bin/windows/contrib/4.3/brotli_1.3.0.zip"   
#>  [72] "bin/windows/contrib/4.3/curl_5.2.1.zip"     
#>  [73] "bin/windows/contrib/4.3/gpg_1.2.9.zip"      
#>  [74] "bin/windows/contrib/4.3/js_1.2.zip"         
#>  [75] "bin/windows/contrib/4.3/jsonlite_1.8.8.zip" 
#>  [76] "bin/windows/contrib/4.3/maketools_1.3.0.zip"
#>  [77] "bin/windows/contrib/4.3/mongolite_2.7.3.zip"
#>  [78] "bin/windows/contrib/4.3/openssl_2.1.1.zip"  
#>  [79] "bin/windows/contrib/4.3/protolite_2.3.0.zip"
#>  [80] "bin/windows/contrib/4.3/rjade_0.1.1.zip"    
#>  [81] "bin/windows/contrib/4.3/sys_3.4.2.zip"      
#>  [82] "bin/windows/contrib/4.3/timestamp.txt"      
#>  [83] "bin/windows/contrib/4.3/webp_1.2.0.zip"     
#>  [84] "bin/windows/contrib/4.3/webutils_1.2.0.zip" 
#>  [85] "src/contrib/PACKAGES"                       
#>  [86] "src/contrib/PACKAGES.gz"                    
#>  [87] "src/contrib/RAppArmor_3.2.4.tar.gz"         
#>  [88] "src/contrib/V8_4.4.2.tar.gz"                
#>  [89] "src/contrib/badgen_1.0.0.tar.gz"            
#>  [90] "src/contrib/base64_2.0.1.tar.gz"            
#>  [91] "src/contrib/bcrypt_1.2.tar.gz"              
#>  [92] "src/contrib/brotli_1.3.0.tar.gz"            
#>  [93] "src/contrib/curl_5.2.1.tar.gz"              
#>  [94] "src/contrib/gpg_1.2.9.tar.gz"               
#>  [95] "src/contrib/js_1.2.tar.gz"                  
#>  [96] "src/contrib/jsonlite_1.8.8.tar.gz"          
#>  [97] "src/contrib/maketools_1.3.0.tar.gz"         
#>  [98] "src/contrib/mongolite_2.7.3.tar.gz"         
#>  [99] "src/contrib/openssl_2.1.1.tar.gz"           
#> [100] "src/contrib/protolite_2.3.0.tar.gz"         
#> [101] "src/contrib/rjade_0.1.1.tar.gz"             
#> [102] "src/contrib/sys_3.4.2.tar.gz"               
#> [103] "src/contrib/timestamp.txt"                  
#> [104] "src/contrib/unix_1.5.7.tar.gz"              
#> [105] "src/contrib/webp_1.2.0.tar.gz"              
#> [106] "src/contrib/webutils_1.2.0.tar.gz"